% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_misc.R, R/snakemake_functions.R
\name{append_md}
\alias{append_md}
\title{Append metadata to DNAm data compilations}
\usage{
append_md(
  files.dname = "recount-methylation-files",
  md.dname = "metadata",
  comp.dname = "compilations"
)

append_md(
  files.dname = "recount-methylation-files",
  md.dname = "metadata",
  comp.dname = "compilations"
)
}
\arguments{
\item{files.dname}{Files dir name for instance ("recount-methylation-files")}

\item{md.dname}{Metadata dir name ("metadata"), contained at files.dname.}

\item{comp.dname}{Compilations dir name ("compilations"), contained at 
files.dname.}

\item{ts}{Timestamp.}

\item{comp.fnv}{Vector of regular expression patterns to filter compilation 
files list.}

\item{mdall.fnstr}{Filename string stem to detect aggregate metadata table at 
md.dname.}

\item{verbose}{Whether to show status messages (TRUE).}
}
\value{
NULL, if all appends successful, otherwise a list of the compilation 
files for which appends weren't successful.

NULL, stores new composite metadata table.
}
\description{
Append metadata to DNAm data compilations. Handles metadata addition for 
either HDF5 or HDF5-SummarizedExperiment compilation files.

Append aggregate metadata to compilation files. This step follows 
formation of the compilation files (e.g. rule `get_rg_compilations`, etc.)
and metadata files (e.g. rule `do_mdmap`, etc.).
}
