% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_misc.R
\name{append_md_h5db}
\alias{append_md_h5db}
\title{Append metadata to HDF5 DNAm data compilation file}
\usage{
append_md_h5db(
  ts,
  mdall,
  comp.fpath,
  id.cname = "gsm",
  dsn = "mdall",
  overwrite = TRUE
)
}
\arguments{
\item{ts}{Timestamp.}

\item{mdall}{Valid metadata table containing column id.cname.}

\item{comp.fpath}{File path to a valid h5 compilation file.}

\item{id.cname}{Sample ID column name in mdall ("gsm").}

\item{dsn}{Name stem of the metadata table in h5db object (e.g. writes table
named dsn and column names to paset0(dsn".colnames") to h5db)}
}
\value{
NULL, saves h5 object with appended metadata.
}
\description{
Append metadata to DNAm HDF5 data compilations.
}
