% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_misc.R
\name{append_md_h5se}
\alias{append_md_h5se}
\title{Append metadata to HDF5-SummarizedExperiment DNAm data compilation file}
\usage{
append_md_h5se(ts, mdall, comp.fpath, id.cname = "gsm", overwrite = TRUE)
}
\arguments{
\item{ts}{Timestamp.}

\item{mdall}{Valid metadata table containing column id.cname.}

\item{comp.fpath}{File path to a valid h5se compilation file.}

\item{id.cname}{Sample ID column name in mdall ("gsm").}

\item{overwrite}{Whether to overwrite existing metadata (TRUE).}
}
\value{
NULL, saves h5se object with appended metadata.
}
\description{
Append metadata to DNAm HDF5-SummarizedExperiment data compilations.
}
