% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{appendvar}
\alias{appendvar}
\title{Append mapped terms to a metadata variable}
\usage{
appendvar(var, val, filtv, m = mdpost)
}
\arguments{
\item{var}{Variable in m for which to append new terms.}

\item{val}{Character string to append to matched samples.}

\item{filtv}{Vector of boolean values identifying samples for which to
append the new term.}

\item{m}{The postprocessed metadata for which to append the new terms
(data.frame, mdpost).}
}
\value{
The result of appending new terms to specified var.
}
\description{
Appends new variable data to a metadata variable, preserving current
variable terms.
}
\seealso{
get_filt, get_pstr_neg, get_pstr
}
