% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snakemake_functions.R
\name{check_h5_database}
\alias{check_h5_database}
\title{Check DNAm data contained in an HDF5 database}
\usage{
check_h5_database(
  comppath = file.path("recount-methylation-files", "compilations")
)
}
\arguments{
\item{comppath}{Path to compilation files for the `recountmethylation` instance.}
}
\value{

}
\description{
Checks DNAm data stored in an HDF5 database. Uses a GEO lookup to grab 
data for data validation.
}
