% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacontrol_format.R
\name{control_biotinstaining}
\alias{control_biotinstaining}
\title{Biotin staining metrics}
\usage{
control_biotinstaining(
  rm,
  rs,
  gs,
  cdf,
  cnames = c("Biotin_red", "Biotin_green"),
  baseline = 3000,
  biotin.baseline = 1,
  verbose = TRUE
)
}
\arguments{
\item{rm}{Matrix of control metric signals.}

\item{rs}{Red signal data (data.frame, columns are probes, rows are 
samples, column names are addresses, rownames are samples/GSM IDs).}

\item{gs}{Green signal data (data.frame, columns are probes, rows are 
samples, column names are addresses, rownames are samples/GSM IDs).}

\item{cdf}{Control probe annotations (data.frame, cols = properties, 
rows = probes).}

\item{cnames}{Vector of control probe column names for the metric.}

\item{baseline}{Baseline measure for signals (integer, 3000).}

\item{biotin.baseline}{Baseline to use for biotin controls (integer, 1).}

\item{verbose}{Whether to show status messages (TRUE).}
}
\value{
Control biotin staining signal.
}
\description{
Two metrics, 1 per color channel
}
\examples{
cdf <- system.file("extdata", "controldf", "cgcontroldf.rda", 
package = "recountmethylation")

}
