% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacontrol_format.R
\name{control_bsconversion}
\alias{control_bsconversion}
\title{Bisulfite conversion I and II}
\usage{
control_bsconversion(
  rm,
  rs,
  gs,
  cdf,
  baseline = 3000,
  cnames = c("Bisulfite_conversion_I_red", "Bisulfite_conversion_I_green",
    "Bisulfite_conversion_II"),
  biotin.baseline = 1,
  verbose = TRUE
)
}
\arguments{
\item{rm}{Matrix of control metric signals.}

\item{rs}{Red signal data (data.frame, columns are probes, rows are 
samples, column names are addresses, rownames are samples/GSM IDs).}

\item{gs}{Green signal data (data.frame, columns are probes, rows are 
samples, column names are addresses, rownames are samples/GSM IDs).}

\item{cdf}{Control probe annotations (data.frame, cols = properties, 
rows = probes).}

\item{baseline}{Baseline measure for signals (integer, 3000).}

\item{cnames}{Vector of control probe column names for the metric.}

\item{biotin.baseline}{Baseline to use for biotin controls (integer, 1).}

\item{verbose}{Whether to show status messages (TRUE).}
}
\value{
Control bisulfite conversion metrics.
}
\description{
Three metrics. For Bisulfite conversion I, calculates one metric from 
controls C1-3 vs. U1-3 in green channel, the other from controls C4-6 vs. 
U4-6 in red channel. For Bisulfite conversion II, calculates one metric.
}
\examples{
cdf <- system.file("extdata", "controldf", "cgcontroldf.rda", 
package = "recountmethylation")
}
