% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacontrol_format.R
\name{control_hybridization}
\alias{control_hybridization}
\title{Hybridization metrics}
\usage{
control_hybridization(
  rm,
  rs,
  gs,
  cdf,
  baseline = 3000,
  cnames = c("Hybridization_medium-vs-high", "Hybridization_low-vs-medium"),
  biotin.baseline = 1,
  verbose = TRUE
)
}
\arguments{
\item{rm}{Matrix of control metric signals.}

\item{rs}{Red signal data (data.frame, columns are probes, rows are 
samples, column names are addresses, rownames are samples/GSM IDs).}

\item{gs}{Green signal data (data.frame, columns are probes, rows are 
samples, column names are addresses, rownames are samples/GSM IDs).}

\item{cdf}{Control probe annotations (data.frame, cols = properties, 
rows = probes).}

\item{baseline}{Baseline measure for signals (integer, 3000).}

\item{cnames}{Vector of control probe column names for the metric.}

\item{biotin.baseline}{Baseline to use for biotin controls (integer, 1).}

\item{verbose}{Whether to show status messages (TRUE).}
}
\value{
Control hybridization metrics.
}
\description{
Two metrics, one comparing high vs. medium level, the other comparing medium 
vs. low level. Both metrics calculated from Green color channel signals.
}
\examples{
cdf <- system.file("extdata", "controldf", "cgcontroldf.rda", 
package = "recountmethylation")
}
