% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacontrol_format.R
\name{control_restoration}
\alias{control_restoration}
\title{Restoration metric}
\usage{
control_restoration(
  bg.addr,
  rm,
  cname = "Restore",
  gs,
  cdf,
  baseline = 3000,
  biotin.baseline = 1,
  verbose = TRUE
)
}
\arguments{
\item{bg.addr}{Probe address of background control probe.}

\item{rm}{Matrix of control metric signals.}

\item{cname}{Column name of the control probe.}

\item{gs}{Green signal data (data.frame, columns are probes, rows are samples, 
column names are addresses, rownames are samples/GSM IDs).}

\item{cdf}{Control probe annotations (data.frame, cols = properties, rows = probes).}

\item{baseline}{Baseline measure for signals (integer, 3000).}

\item{biotin.baseline}{Baseline to use for biotin controls (integer, 1).}

\item{verbose}{Whether to show status messages (TRUE).}
}
\value{
Control restoration signal.
}
\description{
Single metric, using grn channel
}
\examples{
cdf <- system.file("extdata", "controldf", "cgcontroldf.rda", 
package = "recountmethylation")
}
