% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacontrol_format.R
\name{control_specificity}
\alias{control_specificity}
\title{Specificity metrics}
\usage{
control_specificity(
  rm,
  rs,
  gs,
  cdf,
  baseline = 3000,
  cnames = c("Specificity_I_red", "Specificity_I_grn", "Specificity_II"),
  biotin.baseline = 1,
  verbose = TRUE
)
}
\arguments{
\item{rm}{Matrix of control metric signals.}

\item{rs}{Red signal data (data.frame, columns are probes, rows are 
samples, column names are addresses, rownames are samples/GSM IDs).}

\item{gs}{Green signal data (data.frame, columns are probes, rows are 
samples, column names are addresses, rownames are samples/GSM IDs).}

\item{cdf}{Control probe annotations (data.frame, cols = properties, 
rows = probes).}

\item{baseline}{Baseline measure for signals (integer, 3000).}

\item{cnames}{Vector of control probe column names for the metric.}

\item{biotin.baseline}{Baseline to use for biotin controls (integer, 1).}

\item{verbose}{Whether to show status messages (TRUE).}
}
\value{
Control specificity metrics.
}
\description{
Three metrics (two for I and one for type II), for each color channel and 
BeadArray probe type.
}
\examples{
cdf <- system.file("extdata", "controldf", "cgcontroldf.rda", 
package = "recountmethylation")
}
