% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacontrol_format.R
\name{control_targetremoval}
\alias{control_targetremoval}
\title{Target removal metrics}
\usage{
control_targetremoval(
  rm,
  rs,
  gs,
  cdf,
  verbose = TRUE,
  cnames = c("Target_removal_1", "Target_removal_2"),
  baseline = 3000,
  biotin.baseline = 1
)
}
\arguments{
\item{rm}{Matrix of control metric signals.}

\item{rs}{Red signal data (data.frame, columns are probes, rows are 
samples, column names are addresses, rownames are samples/GSM IDs).}

\item{gs}{Green signal data (data.frame, columns are probes, rows are 
samples, column names are addresses, rownames are samples/GSM IDs).}

\item{cdf}{Control probe annotations (data.frame, cols = properties, 
rows = probes).}

\item{verbose}{Whether to show status messages (TRUE).}

\item{cnames}{Vector of control probe column names for the metric.}

\item{baseline}{Baseline measure for signals (integer, 3000).}

\item{biotin.baseline}{Baseline to use for biotin controls (integer, 1).}
}
\value{
Control target removal metrics.
}
\description{
Two metrics, one for each BeadArray probe type, each calculated from green
 color channel signals.
}
\examples{
cdf <- system.file("extdata", "controldf", "cgcontroldf.rda", 
package = "recountmethylation")
}
