% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpipeline.R
\name{data_mdpost}
\alias{data_mdpost}
\title{Retrieve samples metadata from HDF5 db.}
\usage{
data_mdpost(dbn, dsn)
}
\arguments{
\item{dbn}{Path to HDF5 database file.}

\item{dsn}{Name or group path to HDF5 dataset containing postprocessed 
metadata.}
}
\value{
Postprocessed metadata as a `data.frame`.
}
\description{
Retrieves sample metadata from an HDF5 database.
}
