% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacontrol_format.R
\name{df_bathresh}
\alias{df_bathresh}
\title{Get minimum thresholds for BeadArray controls}
\usage{
df_bathresh()
}
\value{
Data frame containing minimum thresholds for BeadArray controls.
}
\description{
Get a table of the minimum thresholds for BeadArray control assessments. 
These are called by get_bathresh() to do the quality assessments (e.g. 
if signal < threshold -> fail, else pass).
}
\seealso{
get_bathresh(); get_bactrl()
}
