% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{do_h5se_merge}
\alias{do_h5se_merge}
\title{Merge 2 h5se objects with all pdata columns}
\usage{
do_h5se_merge(h5se1.fpath, h5se2.fpath, new.h5se.dpath, h5se.merge.fname)
}
\arguments{
\item{h5se1.fpath}{Path to the first h5se object.}

\item{h5se2.fpath}{Path to the second h5se object.}

\item{new.h5se.dpath}{Path to the directory to contain the merged h5se object.}

\item{h5se.merge.fname}{Name of the new merged h5se object.}
}
\value{
Null, saves new merged h5se object.
}
\description{
Performs a merge on 2 h5se objects, retaining all unique pdata columns
between them.
}
