% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpipeline.R
\name{dt_makefiles}
\alias{dt_makefiles}
\title{Make and manage data.table files for red and grn signal}
\usage{
dt_makefiles(
  gpath,
  idatspath,
  destpath,
  version,
  nts,
  min.cols = 1052641,
  overwrite = TRUE,
  fnstem = "mdat.compilation",
  sepval = " ",
  verbose = TRUE
)
}
\arguments{
\item{gpath}{Path to GSM IDAT files, used for basenames when reading in data.}

\item{idatspath}{Path to idat files to read}

\item{destpath}{Destination path to new signal data tables.}

\item{version}{File version information for file names.}

\item{nts}{NTP timestamp integer, for filenames (see get_metadata function).}

\item{min.cols}{Minimum columns to write DNAm data (integer, ).}

\item{overwrite}{Whether to overwrite existing data.table files with same 
destpath (default TRUE).}

\item{fnstem}{Filename stem for data tables.}

\item{sepval}{Separator symbol for tables (" ").}

\item{verbose}{Whether to return verbose notifications.}
}
\value{
list containing dtcond (try conditions results), and new dt paths 
(reds.path and grns.path)
}
\description{
Handles options for signal tables, returns paths used to write data chunks.
}
