% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpipeline.R
\name{dt_write_rg}
\alias{dt_write_rg}
\title{Write red and grn signal data to flat tables.}
\usage{
dt_write_rg(
  probesv,
  hlinkv,
  idatspath,
  reds.path,
  grns.path,
  sepval = " ",
  verbose = TRUE
)
}
\arguments{
\item{probesv}{Vector of probe labels.}

\item{hlinkv}{Vector of GSM IDAT hlink file names, or file basenames.}

\item{idatspath}{Path to IDAT files directory.}

\item{reds.path}{Path to new red signal data table.}

\item{grns.path}{Path to new grn signal data table.}

\item{sepval}{Separator symbol for data tables.}

\item{verbose}{Whether to include verbose messages.}

\item{min.cols}{Minimum data columns to write DNAm data (integer, )}
}
\value{
NULL, writes data chunks as side effect
}
\description{
This function takes a filtered list of valid IDAT hlink files and 
attempts to read their data in chunks with `minfi::read.metharray()`. 
To read as many samples as possible, the function attempts to read 
and combine individual valid samples if an error is thrown for a 
given chunk or sample subset.
}
