% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpipeline.R
\name{dtables_fromsignal}
\alias{dtables_fromsignal}
\title{Make data tables from IDATs}
\usage{
dtables_fromsignal(
  version,
  verbose = TRUE,
  gsmint = 60,
  fnstem = "mdat.compilation",
  sepval = " ",
  getnb = FALSE,
  idats.path = paste("recount-methylation-files", "idats", sep = "/"),
  dest.path = paste("recount-methylation-analysis", "files", "mdata", "compilations",
    sep = "/")
)
}
\arguments{
\item{version}{Version of the run for data table filenames.}

\item{verbose}{Whether to return verbose notifications.}

\item{gsmint}{Number of GSMs to process at a time, typically runs best near 50 samples.}

\item{fnstem}{Filename stem for data tables.}

\item{sepval}{Separator symbol for data being written.}

\item{getnb}{Whether to get noob-normalized Beta-vlaues (default: FALSE).}
}
\value{
Creates new HDF5 database from DNAm signal data.
}
\description{
Make data tables from IDATs, including red and green channel signals and noob-normalized Beta-values.
This function should be run from the "recount-methylation" base directory.
}
