% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snakemake_functions.R
\name{get_all_md}
\alias{get_all_md}
\title{Get available metadata as a single table}
\usage{
get_all_md(
  files.dname = "recount-methylation-files",
  md.dname = "metadata",
  comp.dname = "compilations"
)
}
\arguments{
\item{files.dname}{Files dir name for instance ("recount-methylation-files")}

\item{md.dname}{Metadata dir name ("metadata"), contained at files.dname.}

\item{comp.dname}{Compilations dir name ("compilations"), contained at 
files.dname.}
}
\value{
NULL, stores new composite metadata table.
}
\description{
Detects available metadata and aggregates tables into a single composite 
table.
}
