% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_misc.R
\name{get_atables}
\alias{get_atables}
\title{Generate study annotation tables}
\usage{
get_atables(
  ts,
  atable.fn = "geo_gse-atables_list",
  eq.dname = "equery",
  md.dname = "metadata",
  json.filt.dname = "gsm_json_filt",
  files.dname = "recount-methylation-files",
  verbose = TRUE
)
}
\arguments{
\item{ts}{Timestamp for the preprocessed metadata table to output 
(integer or character).}

\item{atable.fn}{Name of the annotation tables file to output 
("geo_gse-atables_list)}

\item{eq.dname}{Name of directory, in files.dname, containing the equery 
files produced by the server ("equery").}

\item{md.dname}{Name of directory, in files.dname, containing the instance
metadata files ("metadata).}

\item{json.filt.dname}{Name of directory, in files.dname, containing the
filtered GSM JSON files (e.g. containing only sample metadata, 
"gsm_json_filt")}

\item{files.dname}{Main recountmethylation instance files directory 
("recount-methylation-files").}

\item{verbose}{Whether to show status messages (TRUE).}
}
\value{
NULL, produces the tables list file.
}
\description{
Generate a list of study annotation tables from filtered sample JSON files.
Sample/GSM JSON files should be filtered to remove study-level information,
retaining just sample-level metadata.
}
\seealso{
md_preprocess(); md_postprocess()
}
