% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacontrol_format.R
\name{get_background}
\alias{get_background}
\title{Get the background address from control probe data}
\usage{
get_background(cdf, verbose = TRUE)
}
\arguments{
\item{cdf}{Control probe information (data.frame).}

\item{verbose}{Whether to show status messages (TRUE).}
}
\description{
Get the background address from control probe data
}
\examples{
cdf <- system.file("extdata", "controldf", "cgcontroldf.rda", 
package = "recountmethylation")
get_background(cdf)
}
