% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snakemake_functions.R
\name{get_data_md}
\alias{get_data_md}
\title{Retrieve existing instance metadata}
\usage{
get_data_md(instdir = "rmp_instance")
}
\arguments{
\item{instdir}{Path to directory to contain all instance metadata.}
}
\value{
Instance metadata as a list containing the version and timestamp.
}
\description{
User dialogue to handle instance metadata previously generated 
with `new_instance_md()`.
}
