% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_eqfpath}
\alias{get_eqfpath}
\title{Get an edirect query file path}
\usage{
get_eqfpath(
  fn.str = "gsequery_filt.*",
  eqdpath = file.path("recount-methylation-files", "equery")
)
}
\arguments{
\item{fn.str}{File name string matching the queried file.}

\item{eqdpath}{Path to directory containing the equery files}
}
\value{
Path to the queried edirect file.
}
\description{
Lookup and return an edirect query file matching pattern `fn.str`.
}
