% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities 2.R, R/utilities.R
\name{get_filt}
\alias{get_filt}
\title{Get the outcome of pattern matching a string with regex patterns}
\usage{
get_filt(
  v,
  m = mdpre,
  filtrel = "|",
  ntfilt = NULL,
  ptfilt = NULL,
  varl = c("gsm_title", "sample_type", "disease_state", "anatomic_location", "misc")
)

get_filt(
  v,
  m = mdpre,
  filtrel = "|",
  ntfilt = NULL,
  ptfilt = NULL,
  varl = c("gsm_title", "sample_type", "disease_state", "anatomic_location", "misc")
)
}
\arguments{
\item{v}{Character string or vector of such strings.}

\item{m}{Preprocessed metadata used for pattern matching/lookups (data.frame, mdpre).}

\item{filtrel}{Logical symbol joining each regex pattern (default "|").}

\item{ntfilt}{Regex pattern corresponding to negative lookup filter (default NULL).}

\item{ptfilt}{Regex pattern corresponding to positive lookup filter (default NULL).}
}
\value{
The result of assessing a regex match on a metadata variable.

The result of assessing a regex match on a metadata variable.
}
\description{
Does pattern matching for regex patterns constructed from a character
string or vector of such character strings.

Does pattern matching for regex patterns constructed from a character
string or vector of such character strings.
}
\seealso{
appendvar

appendvar
}
