% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gsm_exclude.R
\name{get_gsm_exclude}
\alias{get_gsm_exclude}
\title{Get GSM IDs to exclude for this instance}
\usage{
get_gsm_exclude(
  md = NULL,
  se.path = NULL,
  fn.str = "gsm_exclude",
  fn.dpath = ".",
  id.gsm.cname = "gsm",
  verbose = TRUE
)
}
\arguments{
\item{md}{Metadata table containing sample ID column (id.gsm.cname, 
e.g. "gsm"). If this is NULL, try loading an se object instead.}

\item{se.path}{Path to a valid HDF5-SummarizedExperiment to check/load.}

\item{fn.str}{Name of sample IDs vector file to save.}

\item{fn.dpath}{Directory path to save samples ID vector.}

\item{id.gsm.cname}{Name of column containing the GSM IDs.}

\item{verbose}{Whether to show verbose status messages.}
}
\value{
Vector of GSM IDs, with option to save.
}
\description{
Get the GSM IDs to exclude from target/download for this instance. The
IDs vector informs an updated edirect query filter object (e.g. file
named "gsequery_filt.*").
}
