% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snakemake_functions.R
\name{get_h5se_rg}
\alias{get_h5se_rg}
\title{Get HDF5-SummarizedExperiment of red/green signals, from h5 rg data}
\usage{
get_h5se_rg(
  files.dpath = "recount-methylation-files",
  comp.dname = "compilations",
  ngsm.block = 50
)
}
\arguments{
\item{files.dpath}{Path to instance files directory.}

\item{comp.dname}{Name of compilations directory.}

\item{ngsm.block}{Number of GSM IDs per processed data block.}
}
\value{
NULL, generates an h5se rg database directory as side effect.
}
\description{
Get HDF5-SummarizedExperiment of red/green signals, from h5 rg data
}
