% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_misc.R
\name{get_jsontitle}
\alias{get_jsontitle}
\title{Get sample titles}
\usage{
get_jsontitle(
  ts,
  json.dname = "gsm_json_filt",
  jsonext.regex = ".*\\\\.json.filt$",
  json.titlesdf.fn = "gsm_jsontitledf",
  md.dname = "metadata",
  files.dname = "recount-methylation-files"
)
}
\arguments{
\item{ts}{Timestamp for the preprocessed metadata table to output 
(integer or character).}

\item{json.dname}{Name of directory, in files.dname, containing the instance 
filtered sample/GSM JSON files ("gsm_json_filt").}

\item{jsonext.regex}{Regex pattern to identify filtered JSON files 
(".*\\.json.filt$").}

\item{json.titlesdf.fn}{}

\item{md.dname}{Name of directory, in files.dname, containing the instance 
metadata files ("metadata).}

\item{files.dname}{Main recountmethylation instance files directory 
("recount-methylation-files").}
}
\value{
NULL, produces a file containing the sample titles.
}
\description{
Mines the sample titles from the instance sample/GSM JSON files. These
are included in the mapped metadata tables and used in term mappings.
}
\seealso{
md_preprocess(); md_postprocess(); get_atables()
}
