% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snakemake_functions.R
\name{get_mddnam}
\alias{get_mddnam}
\title{Get DNAm-based metadata}
\usage{
get_mddnam(
  files.dname = "recount-methylation-files",
  md.dname = "metadata",
  comp.dname = "compilations"
)
}
\arguments{
\item{files.dname}{Files dir name for instance ("recount-methylation-files")}

\item{md.dname}{Metadata dir name ("metadata"), contained at files.dname.}

\item{comp.dname}{Compilations dir name ("compilations"), contained at 
files.dname.}
}
\value{
NULL, stores new DNAm-based metadata file at md.dpath.
}
\description{
Gets DNAm-based metadata, including model-based predictions and quality 
metrics.
}
