% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snakemake_functions.R
\name{get_mdmap}
\alias{get_mdmap}
\title{Get mapped metadata}
\usage{
get_mdmap(files.dname = "recount-methylation-files", md.dname = "metadata")
}
\arguments{
\item{files.dname}{Files dir name for instance ("recount-methylation-files")}

\item{md.dname}{Metadata dir name ("metadata").}
}
\value{
NULL, stores new mapped metadata.
}
\description{
Snakemake function to get mapped metadata for a recountmethylation instance.
}
