% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpipeline.R
\name{get_metadata}
\alias{get_metadata}
\title{Append metadata to HDF5 or SE object}
\usage{
get_metadata(title, version, pname = "rmpipeline", sname = "get_timestamp.py")
}
\arguments{
\item{title}{Object title}

\item{version}{Numeric version to be passed, should conform to ##.##.## 
nomenclature}

\item{pname}{Name of pipeline package (default: "rmpipeline")}

\item{sname}{Name of Python script (default: "get_timestamp.y")}
}
\value{
Metadata content for the object
}
\description{
Passes version info and timestamp from Python to object metadata
}
