% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_msrap.R
\name{get_msrap}
\alias{get_msrap}
\title{Get the pipeline data files as a flat table}
\usage{
get_msrap(
  ts,
  files.dir = "recount-methylation-files",
  md.dname = "metadata",
  msrap.dname = "gsm_msrap_outfiles",
  msrap.regex.str = "^msrapout.*",
  gsmid.fnindex = 2,
  gmap.fn = "md_msrapout"
)
}
\arguments{
\item{ts}{NTP timestamp for the new metadata table (string).}

\item{files.dir}{Files dir for the instance ("recount-methylation-files").}

\item{md.dname}{Metadata dir name, located in the files.dir ("metadata").}

\item{msrap.dname}{Name of dir, located in the files.dir, containing 
pipeline outputs for samples/GSM IDs ("gsm_msrap_outfiles").}

\item{msrap.regex.str}{Regex pattern string for valid metadata pipeline 
output files ("^msrapout.*").}

\item{gsmid.fnindex}{File name index of the GSM ID for valid metadata 
pipeline output files (2).}

\item{gmap.fn}{File name stem of the new mapped metadata table 
("md_msrapout").}
}
\value{
NULL, outputs the mapped data table to the metadata directory.
}
\description{
Reads data in the sample/GSM-level metadata files produced by the pipeline 
and "map_msrap.py" script. Data is read and stored as a flat table ready to
be appended to other metadata for the instance.
}
