% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities 2.R, R/utilities.R
\name{get_pstr}
\alias{get_pstr}
\title{Get regex patterns for a string, or the affirmative match for the 
given string}
\usage{
get_pstr(v)

get_pstr(v)
}
\arguments{
\item{v}{Character string or vector of such character strings.}
}
\value{
Regex patterns for detecting matching strings from metadata.

Regex patterns for detecting matching strings from metadata.
}
\description{
Gets the regex pattern matching the affirmative of a given string `v`.
Does progressive capitalization on values separated by spaces
for each value in v, appends flanking '.*' (matches any char)
for each value in v, appends "|" OR conditional separator

Gets the regex pattern matching the affirmative of a given string `v`.
Does progressive capitalization on values separated by spaces
for each value in v, appends flanking '.*' (matches any char)
for each value in v, appends "|" OR conditional separator
}
\seealso{
get_pstr_neg, get_filt, appendvar

get_pstr_neg, get_filt, appendvar
}
