% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities 2.R, R/utilities.R
\name{get_pstr_neg}
\alias{get_pstr_neg}
\title{Get negative regex string mapping}
\usage{
get_pstr_neg(pstr)

get_pstr_neg(pstr)

get_pstr_neg(pstr)

get_pstr_neg(pstr)
}
\arguments{
\item{pstr}{Regex patterns for matching, or output from `get_pstr()`.}
}
\value{
Mapped negative lookup strings.

Regex for negation of indicated patterns.

Mapped negative lookup strings.

Regex for negation of indicated patterns.
}
\description{
Do negative regex term lookup using patterns mapped by pstr().

Does progressive capitalization on values separated by spaces
for each value in v, appends flanking '.*' (matches any char)
for each value in v, appends "|" OR conditional separator

Do negative regex term lookup using patterns mapped by pstr().

Does progressive capitalization on values separated by spaces
for each value in v, appends flanking '.*' (matches any char)
for each value in v, appends "|" OR conditional separator
}
\seealso{
pstr, get_filt, appendvar

get_pstr, get_filt, appendvar

pstr, get_filt, appendvar

get_pstr, get_filt, appendvar
}
