% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_dnam.R
\name{get_qcmetrics}
\alias{get_qcmetrics}
\title{Get quality metrics from DNAm assays}
\usage{
get_qcmetrics(
  ts,
  rgset.fname,
  gmset.fname,
  mdqc.fname = "mdqc",
  athresh = 0.1,
  nsamp.block = 50,
  md.dname = "metadata",
  comp.dname = "compilations",
  files.dname = "recount-methylation-files",
  verbose = TRUE
)
}
\arguments{
\item{ts}{Timestamp for the preprocessed metadata table to output 
(integer or character).}

\item{mdqc.fname}{Name of the quality metrics table output ("mdqc")}

\item{athresh}{Similarity threshold (percent similarity) for the ewastools 
predicted genotype (decimal, 0.1).}

\item{nsamp.block}{Samples per data block processed (integer, 50).}

\item{md.dname}{Name of directory, in files.dname, containing the instance
metadata files ("metadata).}

\item{files.dname}{Main recountmethylation instance files directory 
("recount-methylation-files").}

\item{verbose}{Whether to show status messages (TRUE).}
}
\value{
NULL, produces a table of quality metrics.
}
\description{
Calculates quality metrics from DNAm assays contained in the compilation 
files. These include signals for 17 BeadArray controls, 2 controls for the 
methylated and unmethylated signals, and predicted genotypes by sample
using methods in the ewastools package.
}
\seealso{
md_predictions(); md_preprocess(); md_postprocess()
}
