% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snakemake_functions.R
\name{get_rg_dtables}
\alias{get_rg_dtables}
\title{Get compiled data tables of red and green signals}
\usage{
get_rg_dtables(
  files.dname = "recount-methylation-files",
  comp.dname = "compilations"
)
}
\arguments{
\item{files.dname}{Name of the instance files directory.}

\item{comp.dname}{Name of the instance compilations directory.}
}
\value{
NULL, generates data tables as side effect
}
\description{
Generates the data tables compiling DNAm red and green signals 
from downloaded, and validated, IDATs. Makes two tables, one for 
each channel type. This is called by the rules `get_rg_compilations` 
and `run_dnam_pipeline`.
}
