% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpipeline.R
\name{getblocks}
\alias{getblocks}
\title{Get list of index blocks}
\usage{
getblocks(slength, bsize)
}
\arguments{
\item{slength}{Total length of index vector}

\item{bsize}{Size of index blocks along length}
}
\value{
List of index blocks of min length `slength`/`bsize`
}
\description{
Get list of index blocks, allowing for remainders.
}
