% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpipeline.R
\name{h5.newtables}
\alias{h5.newtables}
\title{Make new minfi data tables from base h5 db}
\usage{
h5.newtables(
  dbn,
  dsn.nb = "noobbeta",
  dsn.meth = "methylated_signal",
  dsn.unmeth = "unmethylated_signal",
  dsn.red = "redsignal",
  dsn.grn = "greensignal",
  verbose = TRUE,
  ngsm.block = 50,
  ncol.chunk = 5000
)
}
\arguments{
\item{dbn}{Name or path to h5 db.}

\item{dsn.nb}{Name of the new noobbeta h5 dataset.}

\item{dsn.meth}{Name of the new methylated h5 dataset.}

\item{dsn.unmeth}{Name of the new unmethylated h5 dataset}

\item{dsn.red}{Name of existing red signal h5 dataset.}

\item{dsn.grn}{Name of existing green signal h5 dataset.}

\item{verbose}{Whether to show verbose status messages.}

\item{ngsm.block}{Number of GSMs (samples) per process block (default 50).}

\item{ncol.chunk}{Number of columns (GSMs/samples) per chunk in saved h5 datasets.}
}
\value{
Adds new minfi h5 datasets to specified h5 db (dbn).
}
\description{
Uses minfi to populate new data tables (noobbeta, methylated and unmethylated signal) from h5 db with red and green signal data.
}
