% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpipeline.R
\name{h5_add_tables}
\alias{h5_add_tables}
\title{Add red or green signal data tables to an HDF5 database.}
\usage{
h5_add_tables(dsn, fnl, platform, fnpath, dbn, ngsm.block, verbose)
}
\arguments{
\item{dsn}{Name of new signal data table.}

\item{fnl}{Vector of signal tables containing data to be added.}

\item{platform}{DNAm array platform for instance.}

\item{fnpath}{Path to dir containing files in fnl.}

\item{dbn}{Database path to write data.}

\item{ngsm.block}{Number of GSMs (samples) per process block (default 50)}

\item{verbose}{Whether to show verbose status updates (TRUE).}
}
\value{
Populates a new HDF5 database with red or green signal data.
}
\description{
Add red or green signal data tables to an HDF5 database.
}
\seealso{
make_h5db_rg
}
