% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpipeline.R
\name{h5_addmd}
\alias{h5_addmd}
\title{Append sample metadata to HDF5 or SE object}
\usage{
h5_addmd(dbn, mdpath, dsn = "mdpost", verbose = TRUE)
}
\arguments{
\item{dbn}{Name of H5 dataset, passed from make_h5db().}

\item{mdpath}{Path to metadata file.}

\item{dsn}{Name of new entity in HDF5 database.}

\item{verbose}{Whether to return verbose status updates.}
}
\value{
Adds metadata table and colnames object to HDF5 database
}
\description{
Add sample metadata table to the HDF5 database.
}
