% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpipeline.R
\name{h5_addtables}
\alias{h5_addtables}
\title{Append metadata to HDF5 or SE object}
\usage{
h5_addtables(dbn, fnl, dsnl, rmax, cmax, verbose = TRUE, nr.inc = 10)
}
\arguments{
\item{dbn}{Name of the HDF5 database targeted.}

\item{fnl}{List of signal tables containing data to be added.}

\item{dsnl}{List of data set names in HDF5 database to be populated.}

\item{rmax}{Total rows to append to data sets, reflecting total samples.}

\item{cmax}{Total columns to append to data sets, reflecting total assays or probes.}

\item{verbose}{Whether to print verbose progress messages.}

\item{nr.inc}{Number of samples to append at a time (default: 10).}
}
\value{
Populates the HDF5 database
}
\description{
Add signal data (red and green channel) to the HDF5 database.
}
