% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpipeline.R
\name{make.h5se}
\alias{make.h5se}
\title{Use DelayedArray function to store H5 SummarizedExperiment directory}
\usage{
make.h5se(
  newfn,
  dsn.data1,
  dsn.data2 = NULL,
  mdpost,
  dsn.rn = "redsignal.rownames",
  se = c("rg", "gr", "gm"),
  verbose = TRUE,
  dsn.cn = FALSE,
  dbn = "remethdb2.h5",
  replace.opt = TRUE
)
}
\arguments{
\item{newfn}{Filename of new H5 SE directory.}

\item{dsn.data1}{Name of first dataset in h5 db (required).}

\item{dsn.data2}{Name of second dataset in h5 db (required if se either rg or gm).}

\item{mdpost}{Sample metadata file (columns are fields, rows are samples).}

\item{dsn.rn}{Name of object in h5 db with the rownames (GSM or sample basenames).}

\item{se}{Type of SummarizedExperiment to parse (either rg for RGChannelSet, gr for GenomicRatioSet, or gm for GenomicMethylSet).}

\item{verbose}{Whether to show verbose status messages.}

\item{dsn.cn}{Name of h5 db object with column names (cgids or assays, optional).}

\item{dbn}{Name or path to h5 db.}

\item{replace.opt}{Whether to replace/overwrite any existing H5 SE directories of the same name as `newfn`.}
}
\value{
List of index blocks of min length `slength`/`bsize`
}
\description{
Use DelayedArray function to store H5 SummarizedExperiment directory from h5 db. Handles 3 classes of Summarized Experiment objects.
}
