% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpipeline.R
\name{make_h5_gm}
\alias{make_h5_gm}
\title{Make an HDF5 db file with MethylSet data, from an RGChannelSet 
h5se file}
\usage{
make_h5_gm(
  dbn,
  version,
  ts,
  blocksize = 65,
  files.dname = "recount-methylation-files",
  comp.dname = "compilations",
  platform = c("hm450k", "epic"),
  newfnstem = "remethdb",
  verbose = TRUE,
  dsv = c("meth", "unmeth"),
  replace.opt = TRUE
)
}
\arguments{
\item{dbn}{Path to the RGChannelSet H5SE data directory.}

\item{version}{Version for new h5 file.}

\item{ts}{Timestamp for new h5 file.}

\item{blocksize}{Size of blocks/number of samples to process at a time 
(default 65).}

\item{files.dname}{Files dir name for instance ("recount-methylation-files")}

\item{comp.dname}{Compilations dir name, located in files.dname}

\item{platform}{Array platform (either "hm450k" or "epic").}

\item{newfnstem}{New filename stem for h5 file.}

\item{verbose}{Whether to return verbose messages (default TRUE).}

\item{dsv}{Vector of new datasets to add to h5 HDF5 database.}

\item{replace.opt}{Whether to replace h5 files with duplicate filename 
(default TRUE).}
}
\value{
Path to new HDF5 object.
}
\description{
Makes an HDF5 db file with raw/unnormalized methylated and unmethylated 
signal, including rows and columns as datasets.
}
