% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpipeline.R
\name{make_h5_gr}
\alias{make_h5_gr}
\title{Make an HDF5 h5 file with GenomicRatioSet data, from an RGChannelSet 
h5se file}
\usage{
make_h5_gr(
  dbn,
  version,
  ts,
  blocksize = 65,
  files.dname = "recount-methylation-files",
  comp.dname = "compilations",
  verbose = TRUE,
  platform = c("hm450k", "epic"),
  newfnstem = "remethdb",
  replace.opt = TRUE
)
}
\arguments{
\item{dbn}{Path to the RGChannelSet H5SE data directory.}

\item{version}{Version for new h5 file.}

\item{ts}{Timestamp for new h5 file.}

\item{blocksize}{Size of blocks/number of samples to process at a time 
(default 65).}

\item{files.dname}{Files dir name for instance ("recount-methylation-files")}

\item{comp.dname}{Compilations dir name, located in files.dname}

\item{verbose}{Whether to return verbose messages (default TRUE).}

\item{platform}{Array platform (either "hm450k" or "epic").}

\item{newfnstem}{New filename stem for h5 file.}

\item{replace.opt}{Whether to replace h5 files with duplicate filename 
(default TRUE).}
}
\value{
Path to new h5 object.
}
\description{
Makes an HDF5 h5 file containing noob-normalized Beta-values, including
rows and columns as separate datasets.
}
