% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpipeline.R
\name{make_h5db}
\alias{make_h5db}
\title{Make and populate a new HDF5 database}
\usage{
make_h5db(
  dbn,
  fnl,
  rmax = 35500,
  cmax = 622399,
  newtables = TRUE,
  dsnl = c("redsignal", "greensignal", "noobbeta")
)
}
\arguments{
\item{dbn}{Name of the HDF5 database targeted.}

\item{fnl}{List of signal tables containing data to be added.}

\item{rmax}{Total rows to append to data sets, reflecting total samples.}

\item{cmax}{Total columns to append to data sets, reflecting total assays or probes.}

\item{newtables}{Whether to also add new data tables (noob-norm. Beta-values, meth. and unmeth. signal).}

\item{dsnl}{List of data set names in HDF5 database to be populated.}
}
\value{
Populates the HDF5 database
}
\description{
Add signal data (red and green channel) to the HDF5 database.
}
