% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpipeline.R
\name{make_h5db_rg}
\alias{make_h5db_rg}
\title{Make and populate a new HDF5 database with red and green signal, and metadata}
\usage{
make_h5db_rg(
  dbfnstem,
  dbpath,
  version,
  ts,
  fnpath,
  platform = c("hm450k", "epic"),
  fnl,
  dsnl = c("redsignal", "greensignal"),
  rmoldh5 = TRUE,
  newtables = FALSE,
  mdpath = NULL,
  ngsm.block = 50,
  verbose = TRUE
)
}
\arguments{
\item{dbfnstem}{Stem of filename for h5 database (to which ts and version 
appended)}

\item{dbpath}{Database file path to write to.}

\item{version}{Version of new database file.}

\item{ts}{Timestamp of new database file.}

\item{fnpath}{Path to dir containing files in fnl.}

\item{platform}{DNAm array platform for instance.}

\item{fnl}{Vector of signal tables containing data to be added.}

\item{dsnl}{Vector of new h5 data set objects in the h5db object, 
corresponding (1:1) to the files declared in fnl.}

\item{newtables}{Whether to also add new data tables (noob-norm. Beta-values, 
meth. and unmeth. signal).}

\item{mdpath}{If addmd, the path to the metadata file to load.}

\item{ngsm.block}{Number of GSMs (samples) per process block (default 50, 
passed to `h5_newtables()`).}

\item{verbose}{Whether to show verbose status updates.}
}
\value{
Populates the HDF5 database
}
\description{
Make and populate a new HDF5 database with red and green signal, and metadata
}
