% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpipeline.R
\name{make_h5se_gm}
\alias{make_h5se_gm}
\title{Make an H5SE MethySet object from a MethySet HDF5 database file}
\usage{
make_h5se_gm(
  dbn,
  version,
  ts,
  platform = c("hm450k", "epic"),
  files.dname = "recount-methylation-files",
  comp.dname = "compilations",
  replaceopt = TRUE,
  verbose = TRUE,
  add.metadata = FALSE,
  pdata = NULL,
  newdnstem = "remethdb",
  semd = list(title = paste0("GenomicMethylSet HDF5-", "SummarizedExperiment"),
    preprocessing = "raw")
)
}
\arguments{
\item{dbn}{Name of the h5 file to read data from.}

\item{version}{Version for new filenames.}

\item{ts}{NTP timestamp.}

\item{platform}{Array platform (either "hm450k" or "epic").}

\item{files.dname}{Files dir name for instance ("recount-methylation-files")}

\item{comp.dname}{Compilations dir name, located in files.dname}

\item{replaceopt}{Whether to overwrite existing file of same name as new file
(default = TRUE).}

\item{verbose}{Whether to return verbose messages (default TRUE).}

\item{add.metadata}{Whether to add samples metadata (boolean, FALSE).}

\item{pdata}{Metadata object obtained from pData(se).}

\item{newdnstem}{Stem for new h5se object file.}
}
\value{
NULL, generates a new h5se file.
}
\description{
Make an H5SE MethySet object from a MethySet HDF5 database file
}
