% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpipeline.R
\name{make_h5se_gr}
\alias{make_h5se_gr}
\title{Make an h5se object from an h5se gr file}
\usage{
make_h5se_gr(
  version,
  ts,
  h5name,
  pdata,
  replaceopt = TRUE,
  newdnstem = "remethdb-h5se"
)
}
\arguments{
\item{version}{Version for new filenames.}

\item{ts}{NTP timestamp.}

\item{h5name}{Name of the h5 file to read data from.}

\item{pdata}{Metadata object obtained from pData(se).}

\item{replaceopt}{Whether to overwrite existing file of same name as new file
(default = TRUE).}

\item{newdnstem}{Stem for new h5se object file.}
}
\value{
NULL, generates a new h5se file.
}
\description{
Make an h5se object from an h5se gr file
}
