% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpipeline.R
\name{makeh5db_rg}
\alias{makeh5db_rg}
\title{Make and populate a new HDF5 database with red and green signal, and metadata}
\usage{
makeh5db_rg(
  dbfnstem,
  version,
  ts,
  fnl,
  fnpath,
  rmax = 35300,
  cmax = 622399,
  rmoldh5 = TRUE,
  newtables = FALSE,
  mdpath = NULL,
  nr.inc = 10,
  ngsm.block = 50,
  verbose = TRUE,
  dsnl = c("redsignal", "greensignal", "mdpost")
)
}
\arguments{
\item{dbfnstem}{Stem of filename for h5 database (to which ts and version appended)}

\item{version}{Version of new database file.}

\item{ts}{Timestamp of new database file.}

\item{fnl}{Vector of signal tables containing data to be added.}

\item{fnpath}{Path to dir containing files in fnl.}

\item{rmax}{Total rows to append to data sets, reflecting total samples.}

\item{cmax}{Total columns to append to data sets, reflecting total assays or probes. Should be 622399 for raw red/grn signal.}

\item{newtables}{Whether to also add new data tables (noob-norm. Beta-values, meth. and unmeth. signal).}

\item{mdpath}{If addmd, the path to the metadata file to load.}

\item{nr.inc}{Number of samples to append at a time (default: 10, passed to `h5_addtables()`).}

\item{ngsm.block}{Number of GSMs (samples) per process block (default 50, passed to `h5_newtables()`).}

\item{verbose}{Whether to show verbose status updates.}

\item{dsnl}{Vector of new h5 data set objects in the h5db object, corresponding (1:1) to the files declared in fnl.}
}
\value{
Populates the HDF5 database
}
\description{
Add signal data (red and green channel) to the HDF5 database.
}
