% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities 2.R, R/utilities.R
\name{match1to2}
\alias{match1to2}
\title{Match and combine 2 datasets}
\usage{
match1to2(d1, d2, ci1 = 2, ci2 = 1)

match1to2(d1, d2, ci1 = 2, ci2 = 1)
}
\arguments{
\item{d1}{First set to combine (matrix).}

\item{d2}{Second set to combine (matrix).}

\item{ci1}{Column index in first dataset to match on}

\item{ci2}{Column index in second dataset to match on}
}
\value{
Union of the 2 datasets, including NA values where appropriate.

Union of the 2 datasets, including NA values where appropriate.
}
\description{
Orders 2 datasets on corresponding variables, then appends an NA
matrix as needed before binding on columns. This is mainly used to 
combine different types of sample metadata on common/shared GSM IDs.
Note columns should be matchable after calling `as.character()` for
each.

Orders 2 datasets on corresponding variables, then appends an NA
matrix as needed before binding on columns. This is mainly used to 
combine different types of sample metadata on common/shared GSM IDs.
Note columns should be matchable after calling `as.character()` for
each.
}
