% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_postprocess.R
\name{md_age_infol}
\alias{md_age_infol}
\title{Age terms to seed regex queries}
\usage{
md_age_infol()
}
\value{
List of terms (names) and string queries (values).
}
\description{
Age terms to seed regex queries, called by md_post_handle_age().
}
