% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_postprocess.R
\name{md_age_unitl}
\alias{md_age_unitl}
\title{Age unit terms to seed regex queries}
\usage{
md_age_unitl()
}
\value{
List of unit term labels (names) and string query seed terms 
(values).
}
\description{
Age unit terms to seed regex queries, called by md_post_handle_age().
}
