% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_misc.R
\name{md_agg}
\alias{md_agg}
\title{Aggregate available metadata tables}
\usage{
md_agg(
  ts,
  platform = NULL,
  lfn = NULL,
  id.cname = "gsm",
  mda.fn = "mdall",
  md.fnv = c("^md_postprocess.*", "^mdmod_dnam-predictions_.*", "^mdqc_.*",
    "^mdrep_.*", "^md_msrapout_.*"),
  md.dpath = file.path("recount-methylation-files", "metadata"),
  verbose = TRUE
)
}
\arguments{
\item{ts}{Timestamp.}

\item{platform}{The DNAm array platform used.}

\item{lfn}{Optional list of filenames. If NULL, automatically detect 
valid metadata files (NULL).}

\item{id.cname}{Column name for sample IDs, which should be present in 
each table to be compiled ("gsm").}

\item{mda.fn}{Filename stem of new aggregate metadata table to save 
("mdall").}

\item{md.fnv}{Vector of regex patterns to detect valid metadata tables.}

\item{md.dpath}{Path to metadata directory.}

\item{verbose}{Whether to show status messages (TRUE).}
}
\value{
NULL, stores aggregate/composite metadata table.
}
\description{
Detects valid metadata at md.dpath, then aggregates these into a single
composite table.
}
