% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_postprocess.R
\name{md_post_cancer}
\alias{md_post_cancer}
\title{Cancer terms for tissue variable}
\usage{
md_post_cancer()
}
\value{
List of terms (names) and string queries (values).
}
\description{
List containing search terms and string queries for regex pattern matching.
}
