% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_postprocess.R
\name{md_post_disease}
\alias{md_post_disease}
\title{Disease terms for disease variable}
\usage{
md_post_disease()
}
\value{
List of terms (names) and string queries (values).
}
\description{
List containing search terms and string queries for regex pattern matching.
}
