% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_postprocess.R
\name{md_post_handle_age}
\alias{md_post_handle_age}
\title{Handle age term mappings for metadata postprocessing}
\usage{
md_post_handle_age(
  mdpre,
  mdpost,
  mdpost.vl = list(age = "age"),
  mdpre.vl = list(sample_id = "gsm", sample_title = "gsm_title", info = "info", age =
    "age", age_temp = "age_temp"),
  verbose = TRUE
)
}
\arguments{
\item{mdpre}{Table of preprocessed metadata.}

\item{mdpost}{Table of postprocessed metadata.}

\item{mdpost.vl}{List mapping term categories (names) to column
names in the mdpost postprocessed metadata matrix to be generated.}

\item{mdpre.vl}{List mapping term categories (names) to column
names in the mdpre preprocessed metadata matrix.}

\item{verbose}{Whether to show status messages (TRUE).}
}
\value{
Postprocessed metadata table with a new column of mapped age
info, values.
}
\description{
This function is called by md_postprocess() in order to handle the logic of
age term mappings from preprocessed metadata.
}
\seealso{
md_postprocess(); md_preprocess()
}
