% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_postprocess.R
\name{md_post_leukemia}
\alias{md_post_leukemia}
\title{Leukemia terms list for disease variable}
\usage{
md_post_leukemia()
}
\value{
List of terms (names) and string queries (values).
}
\description{
List containing search terms and string queries for regex pattern matching.
}
