% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_postprocess.R
\name{md_post_tissue}
\alias{md_post_tissue}
\title{Non-blood terms list for tissue variable}
\usage{
md_post_tissue()
}
\value{
List of terms (names) and string queries (values).
}
\description{
List containing search terms and string queries for regex pattern matching.
}
