% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_postprocess.R
\name{md_postprocess}
\alias{md_postprocess}
\title{Postprocess metadata prepared using md_preprocess()}
\usage{
md_postprocess(
  ts,
  mdpre,
  mdpost.fname = "md_postprocess",
  md.dpath = file.path("recount-methylation-files", "metadata"),
  mdpre.vl = list(study_id = "gse", sample_id = "gsm", sample_title = "gsm_title",
    disease = "disease_state", sample_type = "sample_type", sex = "sex", info = "info",
    age = "age", age_temp = "age_temp"),
  mdpost.vl = list(tissue = "tissue", disease = "disease", age = "age", sex = "sex",
    storageinfo = "storageinfo"),
  disease.search.vars = c("sample_title", "disease"),
  tissue.search.vars = c("sample_type", "sample_title"),
  storage.info.vars = c("sample_type", "sample_title", "info"),
  verbose = TRUE
)
}
\arguments{
\item{ts}{The timestamp for this run.}

\item{mdpre}{The matrix containing preprocessed metadata (returned from
md_preprocess()).}

\item{mdpost.fname}{Filename for newly mapped postprocessed metadata.}

\item{md.dpath}{Path to the directory containing the preprocessed 
metadata matrix, where newly postprocessed metadata will be stored.}

\item{mdpre.vl}{List mapping term categories (names) to column
names in the mdpre preprocessed metadata matrix.}

\item{mdpost.vl}{List mapping term categories (names) to column
names in the mdpost postprocessed metadata matrix to be generated.}

\item{disease.search.vars}{Term categories to search in preprocessed 
metadata for the disease term mappings}

\item{tissue.search.vars}{Term categories in preprocessed metadata to
search for tissue term mappings.}

\item{storage.info.vars}{Term categories in preprocessed metadata to
search for storage information term mappings.}

\item{verbose}{Whether to show status messages (TRUE).}
}
\value{
Postprocessed metadata table.
}
\description{
Perform postprocessing of previously prepreocessed sample metadata. This 
produces the new harmonized variables (specified by arg mdpost.vl),
where harmonization means terms are mapped, lowercase, and "_" separated.
Variable entries can include multiple terms separated by ";". The args
mdpre.vl and mdpost.vl specify the various variable titles in
the preprocess and postprocess dataset. The vars disease.search.vars,
tissue.search.vars, and storage.info.vars specify the mdpre variables to
search for disease, tissue, and storage info mappings.
}
\seealso{
md_preprocess()
}
