% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_preprocess.R
\name{md_preprocess}
\alias{md_preprocess}
\title{Preprocess sample metadata}
\usage{
md_preprocess(
  ts,
  mdpre.fn = "md_preprocess",
  md.dname = "metadata",
  titlesfn.str = "gsm_jsontitledf",
  atablefn.str = "geo_gse-atables_list",
  files.dname = "recount-methylation-files",
  verbose = TRUE
)
}
\arguments{
\item{ts}{Timestamp for the preprocessed metadata table to output 
(integer or character).}

\item{mdpre.fn}{Name of preprocessed metadata table file to output 
("md_preprocess").}

\item{md.dname}{Name of directory, in files.dname, containing the instance 
metadata files ("metadata).}

\item{titlesfn.str}{Name of sample/GSM titles file ("gsm_jsontitledf).}

\item{atablefn.str}{Name of study annotation tables file 
("geo_gse-atables_list")}

\item{files.dname}{Main recountmethylation instance files directory 
("recount-methylation-files").}

\item{verbose}{Whether to show status messages (TRUE).}
}
\value{
NULL, produces the mdpre preprocessed metadata table.
}
\description{
Preprocess sample metadata by coercing JSON-formatted metadata files
into a flat matrix, with regex pattern matching to identify and 
categorize various variable types. Additional data such as the sample
titles contained in the file declared by the titles.fn argument.
}
\seealso{
md_postprocess()
}
