% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_preprocess.R
\name{mdpre_vars}
\alias{mdpre_vars}
\title{Get the regex patterns for variable mappings}
\usage{
mdpre_vars()
}
\value{
Returns list containing regex patterns for each mapped variable.
}
\description{
Gets the regex patterns from vectors of seed terms for each variable mapped 
by md_preprocess().
}
\seealso{
md_preprocess(); md_postprocess()
}
