% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_preprocess.R
\name{mdpre_vars_disease}
\alias{mdpre_vars_disease}
\title{Terms vector for variable disease}
\usage{
mdpre_vars_disease()
}
\value{
Terms vector
}
\description{
Terms to seed regex pattern matching for md_preprocess().
}
\seealso{
md_preprocess(); md_postprocess()
}
