% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_preprocess.R
\name{mdpre_vars_sex}
\alias{mdpre_vars_sex}
\title{Terms vector for variable sex}
\usage{
mdpre_vars_sex()
}
\value{
Terms vector
}
\description{
Terms to seed regex pattern matching for md_preprocess().
}
\seealso{
md_preprocess(); md_postprocess()
}
