% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_preprocess.R
\name{mdpre_vars_tissue}
\alias{mdpre_vars_tissue}
\title{Terms vector for variable tissue}
\usage{
mdpre_vars_tissue()
}
\value{
Terms vector
}
\description{
Terms to seed regex pattern matching for md_preprocess().
}
\seealso{
md_preprocess(); md_postprocess()
}
