% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snakemake_functions.R
\name{new_instance_md}
\alias{new_instance_md}
\title{Get new instance metadata}
\usage{
new_instance_md(instdir = file.path("recount-methylation-files", "metadata"))
}
\arguments{
\item{instdir}{Path to directory to contain all instance metadata.}
}
\value{
NULL, produces the instance metadata file as side effect.
}
\description{
Generates the new metadata, including version and timestamp, for the
`recountmethylation` instance. This is called by rule `new_instance_md`.
}
