% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities 2.R, R/utilities.R
\name{rmp_handle_metadata}
\alias{rmp_handle_metadata}
\title{Get instance metadata info}
\usage{
rmp_handle_metadata(
  files.dname = "recount-methylation-files",
  md.dname = "metadata"
)

rmp_handle_metadata(
  files.dname = "recount-methylation-files",
  md.dname = "metadata"
)
}
\arguments{
\item{files.dname}{Instance files dir name ("recount-methylation-files").}

\item{md.dname}{Metadata dir name, located in files.dname ("metadata").}
}
\value{
Metadata object containing the instance version and timestamp

Metadata object containing the instance version and timestamp
}
\description{
Looks up and loads previously generated instance metadata 
(e.g. using rule `new_instance_md`), which includes the version and 
timestamp for the `recountmethylation` instance.

Looks up and loads previously generated instance metadata 
(e.g. using rule `new_instance_md`), which includes the version and 
timestamp for the `recountmethylation` instance.
}
