% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities 2.R, R/utilities.R
\name{rmp_handle_platform}
\alias{rmp_handle_platform}
\title{Get instance platform info}
\usage{
rmp_handle_platform(
  settings.fname = "settings.py",
  src.path = file.path("recountmethylation_server", "src")
)

rmp_handle_platform(
  settings.fname = "settings.py",
  src.path = file.path("recountmethylation_server", "src")
)
}
\arguments{
\item{settings.fname}{Name of the settings script containing platform info.}

\item{src.path}{Path the `src` dir in the `recountmethylation_server` repo.}
}
\value{
Platform info as list contianing the accession ID and name.

Platform info as list contianing the accession ID and name.
}
\description{
Retrieves the previously established instance platform info (e.g. using 
rule `set_acc`) from the settings script `settings.py` .

Retrieves the previously established instance platform info (e.g. using 
rule `set_acc`) from the settings script `settings.py` .
}
