% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpipeline.R
\name{se.addpheno}
\alias{se.addpheno}
\title{Append phenotype data to a SummarizedExperiment object}
\usage{
se.addpheno(mdp, se)
}
\arguments{
\item{mdp}{Metadata object (columns are fields, rows are GSMs/samples).}

\item{se}{SummarizedExperiment object.}
}
\value{
SummarizedExperiment object with appended phenotype data.
}
\description{
Append phenotype data to a SummarizedExperiment object
}
