% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpipeline.R
\name{dt_checkidat}
\alias{dt_checkidat}
\title{Validate idats at path, returns gsmu and gpath}
\usage{
dt_checkidat(idatspath, verbose = TRUE)
}
\arguments{
\item{idatspath}{Path to idat files to read}
}
\value{
list containing gsmu and gpath
}
\description{
Checks for valid idat file pairs, checks for latest timestamps, 
files with matching timestamps, and valid red/grn IDAT file pairs, 
gets valid gsms idats dir.
}
