% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpipeline.R
\name{dt_write_rg}
\alias{dt_write_rg}
\title{Writes chunk of red and grn signal data}
\usage{
dt_write_rg(
  gi,
  idatspath,
  gpath,
  reds.path,
  grns.path,
  num.assays = 1052641,
  sepval = " ",
  verbose = TRUE
)
}
\arguments{
\item{gi}{Vector of valid sample basenames to read (e.g. have valid IDAT 
pairs).}

\item{idatspath}{Path to idat files directory.}

\item{reds.path}{Path to new red signal data table.}

\item{grns.path}{Path to new grn signal data table.}

\item{num.assays}{Number of assays to expect before writing (1052641 for 
EPIC).}

\item{sepval}{Separator symbol for data tables.}

\item{verbose}{Whether to include verbose messages.}

\item{min.cols}{Minimum data columns to write DNAm data (integer, )}
}
\value{
NULL, writes data chunks as side effect
}
\description{
Writes chunk of red and grn signal data
}
