% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpipeline.R, R/utilities 2.R, R/utilities.R
\name{get_metadata}
\alias{get_metadata}
\title{Append metadata to HDF5 or SE object}
\usage{
get_metadata(
  title,
  version,
  ts = NULL,
  pname = "rmpipeline",
  sname = "get_timestamp.py"
)

get_metadata(
  title,
  version,
  ts = NULL,
  pname = "rmpipeline",
  sname = "get_timestamp.py"
)

get_metadata(
  title,
  version,
  ts = NULL,
  pname = "rmpipeline",
  sname = "get_timestamp.py"
)
}
\arguments{
\item{title}{Object title}

\item{version}{Numeric version to be passed, should conform to ##.##.## 
nomenclature}

\item{ts}{Timestamp. If NULL, get new timestamp using Python function.}

\item{pname}{Name of pipeline package (default: "rmpipeline")}

\item{sname}{Name of Python script (default: "get_timestamp.y")}
}
\value{
Metadata content for the object

Metadata content for the object

Metadata content for the object
}
\description{
Passes version info and timestamp from Python to object metadata

Passes version info and timestamp from Python to object metadata

Passes version info and timestamp from Python to object metadata
}
