% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpipeline.R
\name{h5_addtables}
\alias{h5_addtables}
\title{Append data tables to HDF5 or SE object}
\usage{
h5_addtables(dbn, fnl, fnpath, dsnl, rmax, cmax, verbose = TRUE, nr.inc = 10)
}
\arguments{
\item{dbn}{Name of H5 dataset, passed from make_h5db().}

\item{fnl}{List of signal data table filenames, corresponding (1:1) to names 
for new h5 datasets in dsnl.}

\item{fnpath}{Path to signal data tables.}

\item{dsnl}{List of data set names in HDF5 database to be populated.}

\item{rmax}{Total rows to append to data sets, reflecting total samples.}

\item{cmax}{Total columns to append to data sets, reflecting total assays or 
probes.}

\item{verbose}{Whether to print verbose progress messages.}

\item{nr.inc}{Number of samples to append at a time (default: 10).}
}
\value{
Populates the HDF5 database
}
\description{
Add signal data (red and green channel) to the HDF5 database.
}
