% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpipeline.R
\name{make_h5_gm}
\alias{make_h5_gm}
\title{Make an HDF5 h5 file with MethylSet data, from an RGChannelSet 
h5se file}
\usage{
make_h5_gm(
  dbn,
  version,
  ts,
  newfnstem = "remethdb_h5",
  verbose = TRUE,
  replace.opt = TRUE,
  blocksize = 65
)
}
\arguments{
\item{dbn}{Path to the RGChannelSet h5se object.}

\item{version}{Version for new h5 file.}

\item{ts}{Timestamp for new h5 file.}

\item{newfnstem}{New filename stem for h5 file.}

\item{verbose}{Whether to return verbose messages (default TRUE).}

\item{replace.opt}{Whether to replace h5 files with duplicate filename 
(default TRUE).}

\item{blocksize}{Size of blocks/number of samples to process at a time 
(default 65).}
}
\value{
Path to new h5 object.
}
\description{
Makes an HDF5 h5 database file with raw/unnormalized methylated and 
unmethylated signal, including rows and columns as datasets.
}
