% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpipeline.R
\name{make_h5db}
\alias{make_h5db}
\title{Make and populate a new HDF5 database}
\usage{
make_h5db(
  dbfnstem,
  version,
  ts,
  fnl,
  fnpath,
  rmax = 35300,
  cmax = 622399,
  rmoldh5 = TRUE,
  newtables = TRUE,
  addmd = TRUE,
  mdpath = NULL,
  dsnl = c("redsignal", "greensignal")
)
}
\arguments{
\item{dbfnstem}{Stem of filename for h5 database (to which ts and version appended)}

\item{version}{Version of new database file.}

\item{ts}{Timestamp of new database file.}

\item{fnl}{Vector of signal tables containing data to be added.}

\item{fnpath}{Path to dir containing files in fnl.}

\item{rmax}{Total rows to append to data sets, reflecting total samples.}

\item{cmax}{Total columns to append to data sets, reflecting total assays or probes. Should be 622399 for raw red/grn signal.}

\item{newtables}{Whether to also add new data tables (noob-norm. Beta-values, meth. and unmeth. signal).}

\item{addmd}{Whether to add metadata to the h5db object.}

\item{mdpath}{If addmd, the path to the metadata file to load.}

\item{dsnl}{Vector of new h5 data set objects in the h5db object, corresponding (1:1) to the files declared in fnl.}
}
\value{
Populates the HDF5 database
}
\description{
Add signal data (red and green channel) to the HDF5 database.
}
