% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpipeline.R
\name{make_h5se_gr}
\alias{make_h5se_gr}
\title{Make an h5se object from an h5se gr file}
\usage{
make_h5se_gr(
  version,
  ts,
  dbn,
  files.dname = "recount-methylation-files",
  comp.dname = "compilations",
  platform = c("hm450k", "epic"),
  replaceopt = TRUE,
  verbose = TRUE,
  add.metadata = FALSE,
  pdata = NULL,
  newdnstem = "remethdb",
  semd = list(title = paste0("GenomicMethylSet HDF5-", "SummarizedExperiment"),
    preprocessing = paste0("Normalization with", " out-of-band ", "signal (noob)"))
)
}
\arguments{
\item{version}{Version for new filenames.}

\item{ts}{NTP timestamp.}

\item{dbn}{Path to HDF5 database file to read data from.}

\item{files.dname}{Files dir name for instance ("recount-methylation-files")}

\item{comp.dname}{Compilations dir name, located in files.dname}

\item{replaceopt}{Whether to overwrite existing file of same name as new file
(default = TRUE).}

\item{verbose}{Whether to return verbose messages (default TRUE).}

\item{add.metadata}{Whether to add samples metadata (boolean, FALSE).}

\item{pdata}{Samples metadata as a DataFrame, path to file, or NULL.}

\item{newdnstem}{Character stme for new h5se object file 
("remethdb_h5se-gr").}
}
\value{
NULL, generates a new h5se file.
}
\description{
Make an h5se object from an h5se gr file
}
