% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpipeline.R
\name{make_h5se_gr}
\alias{make_h5se_gr}
\title{Make an h5se object from an h5se gr file}
\usage{
make_h5se_gr(
  dbn,
  version,
  ts,
  platform = c("hm450k", "epic"),
  replaceopt = TRUE,
  verbose = TRUE,
  add.metadata = FALSE,
  pdata = NULL,
  newdnstem = "remethdb_h5se-gr",
  semd = list(title = "GenomicMethylSet HDF5-SummarizedExperiment object",
    preprocessing = "Normalization with out-of-band signal (noob)")
)
}
\arguments{
\item{dbn}{Path to HDF5 database file to read data from.}

\item{version}{Version for new filenames.}

\item{ts}{NTP timestamp.}

\item{replaceopt}{Whether to overwrite existing file of same name as new file
(default = TRUE).}

\item{verbose}{Whether to return verbose messages (default TRUE).}

\item{add.metadata}{Whether to add samples metadata (boolean, FALSE).}

\item{pdata}{Samples metadata as a DataFrame, path to file, or NULL.}

\item{newdnstem}{Character stme for new h5se object file 
("remethdb_h5se-gr").}
}
\value{
NULL, generates a new h5se file.
}
\description{
Make an h5se object from an h5se gr file
}
