% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpipeline.R
\name{make_h5se_rg}
\alias{make_h5se_rg}
\title{make_h5se_rg}
\usage{
make_h5se_rg(
  version = "0.0.1",
  ts = 1590090412,
  dbn = "remethdb_1590090412_0-0-1.h5",
  newfnstem = "remethdb_h5se-rg",
  dsnv = c("redsignal", "greensignal"),
  dsn.md = "mdpost",
  dsn.md.cn = paste0(dsn.md, ".colnames"),
  verbose = TRUE,
  addmd = TRUE,
  mdpath = NULL,
  replace.opt = TRUE,
  dsn.rnv = c(paste0(dsnv[1], ".rownames"), paste0(dsnv[2], ".rownames")),
  dsn.cnv = c(paste0(dsnv[1], ".colnames"), paste0(dsnv[2], ".colnames")),
  semd = list(title = "Recount methylation H5-SE object", version = version, timestamp
    = ts, preprocessing = "raw")
)
}
\arguments{
\item{version}{The data file version.}

\item{ts}{Data file NTP timestamp (integer).}
}
\value{
New h5se object.
}
\description{
make_h5se_rg
}
