% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rmp_handle_metadata}
\alias{rmp_handle_metadata}
\title{Get instance metadata info}
\usage{
rmp_handle_metadata(
  files.dname = "recount-methylation-files",
  md.dname = "metadata"
)
}
\value{
Metadata object containing the instance version and timestamp
}
\description{
Looks up and loads previously generated instance metadata 
(e.g. using rule `new_instance_md`), which includes the version and 
timestamp for the `recountmethylation` instance.
}
