% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpipeline.R
\name{se_addpheno}
\alias{se_addpheno}
\title{Append phenotype data to a SummarizedExperiment object}
\usage{
se_addpheno(phenopath, se, pdat = NULL, verbose = TRUE)
}
\arguments{
\item{verbose}{Whether to show verbose status messages (default TRUE).}

\item{version}{The data file version}

\item{ts}{Data file NTP timestamp (integer)}

\item{dbn}{Name of h5 file to load data matrices from.}

\item{fnstem}{File name stem for new H5SE file}

\item{dsnv}{Vector of the data set file names to read from h5 file.}

\item{dsn.md}{Name of metadata file in h5 file.}

\item{addmd}{Whether to add sample metadata.}

\item{mdpath}{External path to sample metadata. If FALSE, looks for dsn.md in the dbn h5 file.}

\item{dsn.rnv}{Vector for rownames datasets for dsnv sets.}

\item{dsn.cnv}{Vector of colnames datasets for dsnv sets.}

\item{semd}{H5SE object metadata information.}
}
\value{
New H5SE object.
}
\description{
Append phenotype data to a SummarizedExperiment object.
Note, this data should be a data.frame with certain data specified by colnames "basename" and "gsm".
}
