% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmpipeline.R
\name{se_addpheno}
\alias{se_addpheno}
\title{Append phenotype data to a SummarizedExperiment object}
\usage{
se_addpheno(phenopath, se, verbose = TRUE)
}
\arguments{
\item{phenopath}{Path to properly formatted pheno data file (see details).}

\item{se}{SummarizedExperiment object.}

\item{verbose}{Whether to display status messages.}
}
\value{
SummarizedExperiment object with appended phenotype data.
}
\description{
Append phenotype data to a SummarizedExperiment object.
Note, this data should be a data.frame with certain data specified by colnames "basename" and "gsm".
}
